backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-decouple==3.8

# Admin
django-import-export==3.3.8
django-admin-rangefilter==0.12.0
django-simple-history==3.5.0

# PDFs (optional; needs OS libs)
reportlab==4.2.2
weasyprint==61.2

# Tasks
celery==5.3.6
redis==5.0.7

# Utils
Pillow==10.4.0
tablib[xls,xlsx]==3.5.0
